Tidy up things by moving types to modes/include
