+++ simulator/src/indexable-simulator.cpp
@@ -2,9 +2,8 @@
 
 #include "default_simulation.h"
 
-#include "src/modes/group_type.hpp"
-#include "src/modes/manager_type.hpp"
-#include "src/modes/mode_type.hpp"
+#include "src/modes/include/group_type.hpp"
+#include "src/modes/include/manager_type.hpp"
 
 #include "src/modes/default/fixed_modes.hpp"
 #include "src/modes/legacy/legacy_modes.hpp"
